added README.
These are the examples from the following book:
 David Vandevoorde, Nicolai M. Josuttis
 C++ Template - The Complete Guide
 Addison-Wesley, 2002
 ISBN 0-201-73484-2

For further informations see:
    http://www.josuttis.com/tmplbook/

I welcome your feedback.
The best way to reach me is by Email:
    tmplbook@josuttis.com

Copyright 2003 by Pearson Education, Inc. David Vandevoorde,
and Nicolai M. Josuttis.  All rights reserved.

Permission to use, copy, modify and distribute this software for personal
and educational use is hereby granted without fee, provided that the above
copyright notice appears in all copies and that both that copyright notice
and this permission notice appear in supporting documentation, and that the
names of Pearson Education or the authors are not used in advertising or
publicity pertaining to distribution of the software without specific,
written prior permission.  Pearson Education and the authors make no
representations about the suitability of this software for any purpose.
It is provided "as is" without express or implied warranty.

PEARSON EDUCATION AND THE AUTHORS DISCLAIM ALL WARRANTIES WITH REGARD
TO THIS SOFTWARE, INCLUDING ALL IMPLIED WARRANTIES OF MERCHANTABILITY
AND FITNESS.
IN NO EVENT SHALL PEARSON EDUCATION OR THE AUTHORS BE LIABLE FOR ANY
SPECIAL, INDIRECT OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES WHATSOEVER
RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN ACTION OF
CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR IN
CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.